requirements.txt
numpy
squarify
seaborn
scipy
matplotlib_venn
PyPDF2